Fix vertex location function return type

+++ header/PerlinNoise.hpp
@@ -13,7 +13,7 @@ class PerlinNoise
         PerlinGrid noiseGrid;
         int octaves;
 
-        std::array<float, 8> findCorners(float x, float y);
+        std::array<int, 8> findCorners(float x, float y);
         float dot(Vector& v1, Vector& v2);
         Vector generateUnitVector(void);
